Created subhierarchy for argparse under testing, and seperated conftest to two
